Rename features/signup folder
