+++ .env
@@ -1,0 +1,1 @@
+API_URL=http://localhost:3001
