+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.7.1",
+    "axios": "^0.25.0",
     "formik": "^2.2.9",
     "grommet": "^2.20.0",
     "next": "12.0.8",
