+++ src/services/api.ts
@@ -1,0 +1,3 @@
+import axios from 'axios';
+
+export const api = axios.create({ baseURL: process.env.API_URL });
